Clean up on Matlab script
Cleaned up the Matlab script to parse the GNURadio data file.  Added
better comments and arranged the code better.  Added a routine to
remove zeros from the file.  This happens because of the valve function
in GRC which does not actually stop recording data, but the flow of
data and puts zeros instead.